Adicionado Play Controller
