+++ ApiRest/Controllers/TheatricalPlayersRefactoringKataController.cs
@@ -11,36 +11,11 @@ namespace TheatricalPlayersRefactoringKata.API.Controllers
     public class TheatricalPlayersRefactoringKataController : ControllerBase
     {
         private ApiDbContext _db;
-        private readonly StatementPrinter _statementPrinter;
+        //private readonly StatementPrinter _statementPrinter;
 
-        public TheatricalPlayersRefactoringKataController(ApiDbContext db, StatementPrinter statementPrinter)
+        public TheatricalPlayersRefactoringKataController(ApiDbContext db)//, StatementPrinter statementPrinter)
         {
-            _db = db;
-            _statementPrinter = statementPrinter;
-        }
-
-        [HttpPost]
-        public IActionResult CreatePlay([FromBody] Play play)
-        {
-            if (play == null)
-            {
-                return BadRequest("Play data is null.");
-            }
-            _db.Plays.Add(play);
-            _db.SaveChanges();
-
-            return CreatedAtAction(nameof(GetPlay), new { id = play.Id }, play);
-        }
-        [HttpGet("{id}")]
-        public IActionResult GetPlay(int id)
-        {
-            var play = _db.Plays.Find(id);
-            if (play == null)
-            {
-                return NotFound();
-            }
-
-            return Ok(play);
+            //_db = db;
         }
     }
 }
